requirements.txt
django
django-jazzmin
Pillow
phonenumbers
django-phonenumber-field
git+https://github.com/dmpayton/django-admin-honeypot.git#egg=django-admin-honeypot